feat:add kanban board component along with cards that can be moved around using the newly written CardsAPI
Add new dependencies with dnd-core and uuid.

Column component encapsulates droppable and is used to visually represent a cards state on the kanban board.

KanbanBoard is a main component on this update which will hold in it all the cards and the columns.

Card component encapsulates draggable and is supposed to reside in a parent column which corresponds to its state.

CardsAPI is the way to interact with the stored cardData.

CardsAPIContext provides the CardsAPI for its children to help interact with the backlog.

ECardState is an enum of all the states any given card can be.

ECardType is an enum of all the types any given card can be.

ICardData is an interface supported by all the cards stored; this interface can be used for intellisense as well as reference while programming with card data.