feat: Add MessageService context, most apps can now show a non blocking message to the user with the function showMessage
Add a new dependency uuid which is used while generating messages.

Add couple of spacing values straight out of bulma as css variables available everywhere.

Added three new site settings: notificationDisplayTime: sets the amount of time a notification is shown to the user.
messageServiceTickInterval: lower the better but performance costly. small values are desired or the progress bar on Message component will appear choppy.
maximumMessagesDisplayed: Automatically pop out the first inserted message if number of messages being handled crosses this number, higher number is performance intensive and will also not look good on the screen.

Added the MessageService component into App.

Add the Message component which will ONLY be instantiated by the message service to show notifications to the user.

Add the MessageService component whose whole job is to provide the showMessage function to all its children via the messageServiceContext

DownloadMessage component now shows a danger notification to the user if trying to open the modal with no boards inside the project.