feat: made a skeleton of how the board component must look and feel like.
No pointer events will be raised on the message box as it is messing with scroll bars on the columns.

Added a new component Board, this will hold all columns which will in turn hold all cards (items).

Added a new component Card which is an abstraction of every item.

Added a new component Column which can be used to arrange the cards in the future.