feat: EditPage nav
Add dependencies on just-capitalize and just-clone to both clone an object and capitalize a string.

EditPage is now shown only when atleast one card is already selected.

Dropdown fires an event 'onSelect' every time a new child is selected, also dropdown now plays nice with the htmlFor attribute of label with the help of labelId prop.

Add utility function to ECardState to convert arbitrary strings of any type to ECardState.