feat: Can edit the board of any specific card from the edit/modify page also can add new cards into the project
Removed the CardAPI, the boardAPI also increasingly looks irrelevant, might get rid of it in a future commit.

Added the create card component onto Board and refactored said component to work with the new selectSet and board api's.

Can create a new card on the current board using the CreateCard component, this will also automatically open the modify page.

Refactored the Card component to work with the new selectSet api and the removal of the CardAPI.

Refactored ColumnPage component to work with the new BoardAPI and also removed CardAPI references from it.

EditPage now works with IEditCardProperties instead of cards.

Can now change the board the card is in from the fields component.

Refactored Nav component in the edit page to work with api changes in the edit page.

Massive change to boardAPI, changed from the authority on board level of the project to basically irrelevant as it can only call project level calls with some fields filled in.

Removed the CardAPI, card level responsibility on a board will be handled by the BoardAPI.

Populated the ProjectAPI to handle and store Project state.

Refactored SelectSet functionality into its own API.

Refactored main nav components to work with the new ProjectAPI.

KanbanDateTime can now be serialized and deserialized to a string using the to and from string api's, also added a new api now to get the date time now.

LabeledGrid component no longer capitalizes everything by default and also the component is named as this seems to be required for react to work properly.

LabeledInput will automatically capitalize its label going forward.

ICard interface improved, now is a combination of properties and uuid interfaces.